feat: Rolling 4-hour windows for panic/FOMO indices (#4)
* feat: Add sentiment-based coloring and clickable posts to Live Feed

- Add sentiment score coloring to post cards (green for bullish, red for bearish)
- Display sentiment score value with color coding in post header
- Make posts clickable to open original URL in new tab
- Update API to JOIN with user_sentiment_scores for actual sentiment data
- Change score schema from int to float for proper sentiment values


* feat: Implement rolling 4-hour windows for sentiment indices

- Replace daily grouping with hourly data points using rolling 4-hour aggregation
- Eliminates midnight UTC reset behavior for smoother, continuous index values
- Add fallback mechanism for dashboard summary: 4h -> 24h -> 7 days
- Update merge_history_data to preserve hourly timestamps
- Apply same rolling window logic to per-source sentiment history

This provides more dynamic and responsive panic/FOMO indices that don't
artificially reset at day boundaries.


---------

+++ crypto_sentiment_crawler/dashboard/schemas.py
@@ -155,7 +155,7 @@ class RecentPost(BaseModel):
     content: Optional[str] = None
     url: Optional[str] = None
     author: Optional[str] = None
-    score: Optional[int] = None
+    score: Optional[float] = None  # Sentiment score (-1 to 1)
     created_at: str
     crawled_at: str
